Minor DB optimisations

+++ pgbot/db.py
@@ -18,6 +18,9 @@ db_channel = None
 # Store "name: pickled data" pairs as cache. Do not store unpickled data
 db_obj_cache = {}
 
+# Optimisation: store per-db bool on whether it got updated or not
+db_changed = {}
+
 
 async def init(db_chan: discord.TextChannel):
     """
@@ -33,6 +36,7 @@ async def init(db_chan: discord.TextChannel):
     async for msg in db_channel.history():
         if msg.attachments:
             db_obj_cache[msg.content] = await msg.attachments[0].read()
+            db_changed[msg.content] = False
 
 
 async def quit():
@@ -44,10 +48,13 @@ async def quit():
         return
 
     async for msg in db_channel.history():
-        if msg.content in db_obj_cache:
+        if msg.content in db_obj_cache and db_changed[msg.content]:
             await msg.delete()
 
     for name, picked in db_obj_cache.items():
+        if not db_changed[name]:
+            continue
+
         with io.BytesIO(picked) as fobj:
             await db_channel.send(name, file=discord.File(fobj))
 
@@ -80,4 +87,18 @@ class DiscordDB:
         """
         Store object in DB
         """
-        db_obj_cache[self.name] = pickle.dumps(obj)
+        dumped = pickle.dumps(obj)
+        if dumped != db_obj_cache.get(self.name):
+            db_obj_cache[self.name] = dumped
+            db_changed[self.name] = True
+
+    def delete(self):
+        """
+        Delete DB, returns whether it was deleted successfully
+        """
+        db_changed[self.name] = True
+        try:
+            db_obj_cache.pop(self.name)
+            return True
+        except KeyError:
+            return False
